+++ src/PhpPresentation/Slide/AbstractBackground.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PhpOffice\PhpPresentation\Slide;
+
+class AbstractBackground
+{
+
+}