+++ src/PhpPresentation/Writer/ODPresentation/Content.php
@@ -1212,6 +1212,19 @@ class Content extends AbstractPart
                     break;
             }
         }
+        $oBackground = $slide->getBackground();
+        if ($oBackground instanceof Slide\AbstractBackground) {
+            $objWriter->writeAttribute('presentation:background-visible', 'true');
+            if ($oBackground instanceof Slide\Background\Color) {
+                $objWriter->writeAttribute('draw:fill', 'solid');
+                $objWriter->writeAttribute('draw:fill-color', '#' . $oBackground->getColor()->getRGB());
+            }
+            if ($oBackground instanceof Slide\Background\Image) {
+                $objWriter->writeAttribute('draw:fill', 'bitmap');
+                $objWriter->writeAttribute('draw:fill-image-name', 'background_'.$incPage);
+                $objWriter->writeAttribute('style:repeat', 'stretch');
+            }
+        }
         $objWriter->endElement();
         // > style:style
         $objWriter->endElement();