+++ src/PhpPresentation/Writer/PowerPoint2007.php
@@ -22,6 +22,7 @@ use PhpOffice\PhpPresentation\PhpPresentation;
 use PhpOffice\PhpPresentation\Shape\Drawing as DrawingShape;
 use PhpOffice\PhpPresentation\Shape\MemoryDrawing as MemoryDrawingShape;
 use PhpOffice\PhpPresentation\Shape\Chart as ChartShape;
+use PhpOffice\PhpPresentation\Slide\Background\Image;
 use PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractLayoutPack;
 use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Chart;
 use PhpOffice\PhpPresentation\Writer\PowerPoint2007\ContentTypes;
@@ -269,11 +270,18 @@ class PowerPoint2007 implements WriterInterface
 
             // Add slides (drawings, ...) and slide relationships (drawings, ...)
             for ($i = 0; $i < $this->presentation->getSlideCount(); ++$i) {
+                $oSlide = $this->presentation->getSlide($i);
                 // Add slide
-                $objZip->addFromString('ppt/slides/_rels/slide' . ($i + 1) . '.xml.rels', $wPartRels->writeSlideRelationships($this->presentation->getSlide($i)));
-                $objZip->addFromString('ppt/slides/slide' . ($i + 1) . '.xml', $wPartSlide->writeSlide($this->presentation->getSlide($i)));
-                if ($this->presentation->getSlide($i)->getNote()->getShapeCollection()->count() > 0) {
-                    $objZip->addFromString('ppt/notesSlides/notesSlide' . ($i + 1) . '.xml', $wPartSlide->writeNote($this->presentation->getSlide($i)->getNote()));
+                $objZip->addFromString('ppt/slides/_rels/slide' . ($i + 1) . '.xml.rels', $wPartRels->writeSlideRelationships($oSlide));
+                $objZip->addFromString('ppt/slides/slide' . ($i + 1) . '.xml', $wPartSlide->writeSlide($oSlide));
+                // Add note slide
+                if ($oSlide->getNote()->getShapeCollection()->count() > 0) {
+                    $objZip->addFromString('ppt/notesSlides/notesSlide' . ($i + 1) . '.xml', $wPartSlide->writeNote($oSlide->getNote()));
+                }
+                // Add background image slide
+                $oBkgImage = $oSlide->getBackground();
+                if ($oBkgImage instanceof Image) {
+                    $objZip->addFromString('ppt/media/'.$oBkgImage->getIndexedFilename($i), file_get_contents($oBkgImage->getPath()));
                 }
             }