+++ CHANGELOG.md
@@ -11,9 +11,11 @@
 
 ### Features
 - ODPresentation Reader/Writer : Name of the slide - @Progi1984 GH-121
+- ODPresentation Reader/Writer : Slide Background Color or Image - @Progi1984 GH-152
 - PowerPoint2007 Reader : Support for Layout Name - @Progi1984 GH-144
 - PowerPoint2007 Reader/Writer : Mark as final - @Progi1984 GH-118
 - PowerPoint2007 Reader/Writer : Set default zoom value for presentation - @Progi1984 GH-122
+- PowerPoint2007 Reader/Writer : Slide Background Color or Image - @Progi1984 GH-152
 
 ## 0.5.0 - 2015-10-08